pip_requirements.txt
Flask==0.10.1
nose==1.3.0
Pillow==2.4.0
PyYAML==3.10
requests==2.2.1
yanc==0.2.3

newman==1.2.0
